Add back ModelingToolkit-based ParameterizedFunctions
ParameterizedFunctions.jl was removed in v6 because SymEngine kept having build issues. ParameterizedFunctions.jl v5 is now ModelingToolkit-based, which fixes a few of its issues but also means that it has no scary binaries, so it shouldn't lead to any installation issues. There is no plan to bring it back throughout the documentation, the plan for the future is still to feature ModelingToolkit more prominently, but this will at least fix issues that come up where people find old versions of the documentation and expect this to work. So this is somewhat a backwards compatibility fix: now, old docs will work again.

Fixes https://github.com/JuliaDiffEq/DiffEqDocs.jl/issues/340