Remove some background packages to greatly improve startup time
Most of these packages were undocumented in terms of the DiffEq docs (https://diffeq.sciml.ai/stable/). A few of them had a page mentioning them and pointing to a separate documentation, so they really weren't a part of the package. The only one that really was somewhat core was ParameterizedFunctions.jl, but it was removed from the documentation in 2019. We're now entering 2022 and that deprecated functionality is the cause of more than half of the dependencies because it pulls in ModelingToolkit and the whole symbolic stack. Going into 2022, SciML now separately documents all of the tools which aren't solvers, so DifferentialEquations.jl is about solvers. These don't fit the bill, they add a ton of dependencies, they add a bunch of startup time, so it's time they go bye bye from here.

Result:

```julia
# Before
julia> @time using DifferentialEquations
  8.694322 seconds (24.77 M allocations: 1.736 GiB, 7.27% gc time, 17.92% compilation time)

# After
julia> @time using DifferentialEquations
  5.761738 seconds (18.00 M allocations: 1.327 GiB, 4.54% gc time, 10.38% compilation time)
```

Not to mention precompilation is no longer the easiest way to heat yourself for the winter. 

Merry Christmas. I'm going to find other ways to improve that startup time as well.

@yingboma is there a good way to warn users? I want to just give one warning for `@ode_def` and say "Warning: `@ode_def` is no longer included with DifferentialEquations.jl. Please directly add ParameterizedFunctions.jl" if someone tries to use it? Most users of DiffEq directly won't be savvy enough to read News.md for this.